@@ -1,13 +1,13 @@
 plugins {
     id 'java'
-    id "org.jetbrains.intellij" version "1.6.0"
+    id "org.jetbrains.intellij" version "1.15.0"
 }
 
 group 'com.mzyupc'
-version 'beta-0.9.0'
+version '1.0.0'
 
-sourceCompatibility = JavaVersion.VERSION_11
-targetCompatibility = JavaVersion.VERSION_11
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
 
 repositories {
     mavenCentral()
@@ -32,38 +32,50 @@ dependencies {
 // See https://github.com/JetBrains/gradle-intellij-plugin/
 // 调试插件的 IDEA 版本
 intellij {
-    version = '2022.1.3'
+    version = '2023.2'
 }
 
 patchPluginXml {
-    // 声明插件适用的版本 221.5921.22(2022.1.3)
-    sinceBuild = '221.5921.22'
+    // 声明插件适用的版本 232.5150.116
+    sinceBuild = '232.5150.116'
     untilBuild = provider { null }
     changeNotes = """
       <div>Due to the change in password persistence in beta-0.9.0, the password will be erased after upgrading, so you need to backup your password before upgrading from beta-0.8.0 and earlier to beta-0.9.0 and later.</div>
       <br>
       <div>由于 beta-0.9.0 版本修改了密码的持久化方式, 升级后密码会被清除, 所以从 beta-0.8.0 及其之前的版本升级到 beta-0.9.0 及其之后版本之前需要备份密码</div>
       <br>
-      
+
+      1.0.0<br>
+      <ul>
+          <li>Solved the issue of slow DB opening with a million data entries</li>
+          <li>Resolved the "EventQueue.isDispatchThread()=false" issue</li>
+          <li>Fixed the NPE issue with "Test Connection"</li>
+          <li>Added a complex confirmation method when flushing the DB</li>
+          <li>Updated deprecated API calls</li>
+          <li>Utilized Java 17</li>
+          <li>Added a logo</li>
+          <li>Removed key pagination, now showing up to 20,000 keys. Use the 'Filter' feature if the desired key is not found in the results.</li>
+      </ul>
+
       beta-0.9.0<br>
       <ul>
           <li>Use `PasswordSafe` to save your password</li>
           <li>Upgrade Jedis version to 4.2.3</li>
           <li>Configuration of Username is supported in Connection Settings Dialog</li>
       </ul>
-      
+
       beta-0.8.0</br>
       <ul>
           <li>update: update deprecated API</li>
           <li>fix: background display issues when using third party themes</li>
           <li>fix: freezing the UI while loading</li>
       </ul>
-      
+
       beta-0.7.0</br>
       <ul>
           <li>update: the Dialog's OK/Cancel buttons are arranged in a different order depending on the operating system.</li>
       </ul>
-      
+
       beta-0.6.0</br>
       <ul>
           <li>new: redis connection right-click menu adds "Info" function</li>
@@ -71,7 +83,7 @@ patchPluginXml {
           <li>update: fix redis console line height</li>
           <li>update: use jackson to serialize json</li>
       </ul>
-      
+
       beta-0.5.0</br>
       <ul>
           <li>new: redis console(The 'RESTORE' command and blocking commands such as 'SUBSCRIBE' are not supported)</li>
@@ -79,29 +91,29 @@ patchPluginXml {
           <li>update: query the number of DBs without using the 'CONFIG GET' command</li>
           <li>update: replace Label with Jlabel</li>
       </ul>
-      
+
       beta-0.3.4</br>
       <ul>
           <li>update: add plugin description</li>
       </ul>
-      
+
       beta-0.3</br>
       <ul>
           <li>update: value格式化</li>
       </ul>
-      
+
       beta-0.2</br>
       <ul>
           <li>fix: 不同项目会出现相同的connection</li>
           <li>update: 支持最早版本到201.8743.12</li>
           <li>update: Key分页Page Size的计算</li>
       </ul>
-      
+
       beta-0.1</br>
       <ul>
           <li>first version</li>
       </ul>
-      
+
       """
 }
 test {
