@@ -1,8 +1,11 @@
 package com.mzyupc.aredis.view;
 
-import com.intellij.icons.AllIcons;
 import com.intellij.ide.CommonActionsManager;
-import com.intellij.openapi.actionSystem.*;
+import com.intellij.openapi.actionSystem.ActionManager;
+import com.intellij.openapi.actionSystem.ActionPlaces;
+import com.intellij.openapi.actionSystem.ActionPopupMenu;
+import com.intellij.openapi.actionSystem.ActionToolbar;
+import com.intellij.openapi.actionSystem.DefaultActionGroup;
 import com.intellij.openapi.application.ReadAction;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.LoadingDecorator;
@@ -12,6 +15,7 @@ import com.intellij.ui.components.JBLabel;
 import com.intellij.ui.components.JBScrollPane;
 import com.intellij.ui.treeStructure.Tree;
 import com.intellij.util.containers.Convertor;
+import com.intellij.util.ui.JBUI;
 import com.mzyupc.aredis.action.AddAction;
 import com.mzyupc.aredis.action.ClearAction;
 import com.mzyupc.aredis.action.DeleteAction;
@@ -21,6 +25,7 @@ import com.mzyupc.aredis.utils.RedisPoolManager;
 import com.mzyupc.aredis.utils.ThreadPoolManager;
 import com.mzyupc.aredis.view.dialog.ConfirmDialog;
 import com.mzyupc.aredis.view.dialog.ErrorDialog;
+import com.mzyupc.aredis.view.dialog.FlushDbConfirmDialog;
 import com.mzyupc.aredis.view.dialog.NewKeyDialog;
 import com.mzyupc.aredis.view.render.KeyTreeCellRenderer;
 import com.mzyupc.aredis.vo.DbInfo;
@@ -34,17 +39,19 @@ import org.jetbrains.annotations.NotNull;
 import redis.clients.jedis.Jedis;
 
 import javax.swing.*;
-import javax.swing.border.EmptyBorder;
-import javax.swing.tree.*;
+import javax.swing.tree.DefaultMutableTreeNode;
+import javax.swing.tree.DefaultTreeModel;
+import javax.swing.tree.MutableTreeNode;
+import javax.swing.tree.TreeNode;
+import javax.swing.tree.TreePath;
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Enumeration;
 import java.util.List;
-import java.util.*;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
 
 import static com.intellij.openapi.ui.DialogWrapper.OK_EXIT_CODE;
 import static com.mzyupc.aredis.utils.JTreeUtil.getTreeExpander;
@@ -57,7 +64,6 @@ import static redis.clients.jedis.params.ScanParams.SCAN_POINTER_START;
  */
 @Getter
 public class KeyTreeDisplayPanel extends JPanel {
-    private static final int DEFAULT_PAGE_SIZE = 10000;
 
     private final Tree keyTree;
     private final LoadingDecorator keyDisplayLoadingDecorator;
@@ -68,7 +74,7 @@ public class KeyTreeDisplayPanel extends JPanel {
     /**
      * 每次查询KEY的数量
      */
-    private final int pageSize = DEFAULT_PAGE_SIZE;
+    private static final int DEFAULT_PAGE_SIZE = 20000;
     private DefaultTreeModel treeModel;
     private JPanel keyDisplayPanel;
     /**
@@ -76,30 +82,28 @@ public class KeyTreeDisplayPanel extends JPanel {
      */
     private DefaultMutableTreeNode flatRootNode;
     private JBLabel pageSizeLabel;
-    private JBLabel pageLabel;
-    private int pageIndex = 1;
-    private List<String> allKeys;
-    private List<String> currentPageKeys = new ArrayList<>();
+    private List<String> keys;
 
-    private Map<Integer, String> pageIndexScanPointerMap = new HashMap<>();
-
-    public KeyTreeDisplayPanel(Project project, ARedisKeyValueDisplayPanel parent, JBSplitter splitterContainer, DbInfo dbInfo, RedisPoolManager redisPoolManager, Consumer<KeyInfo> doubleClickKeyAction) {
+    public KeyTreeDisplayPanel(Project project,
+                               ARedisKeyValueDisplayPanel parent,
+                               JBSplitter splitterContainer,
+                               DbInfo dbInfo,
+                               RedisPoolManager redisPoolManager,
+                               Consumer<KeyInfo> doubleClickKeyAction) {
         this.project = project;
         this.dbInfo = dbInfo;
         this.redisPoolManager = redisPoolManager;
         this.parent = parent;
-        pageIndexScanPointerMap.put(pageIndex, SCAN_POINTER_START);
 
-        allKeys = redisPoolManager.scan(SCAN_POINTER_START, parent.getKeyFilter(), pageSize, dbInfo.getIndex());
-
+        keys = redisPoolManager.scanKeys(SCAN_POINTER_START, parent.getKeyFilter(), DEFAULT_PAGE_SIZE, dbInfo.getIndex());
         // exception occurred
-        if (allKeys == null) {
+        if (keys == null) {
             throw new RuntimeException("exception occurred");
         }
 
         keyTree = new Tree();
         // 搜索功能
-        new TreeSpeedSearch(keyTree, new Convertor<TreePath, String>() {
+        TreeSpeedSearch.installOn(keyTree, true, new Convertor<TreePath, String>() {
             @Override
             public String convert(final TreePath treePath) {
                 DefaultMutableTreeNode node = (DefaultMutableTreeNode) treePath.getLastPathComponent();
@@ -108,7 +112,7 @@ public class KeyTreeDisplayPanel extends JPanel {
                 }
                 return "";
             }
-        }, true);
+        });
 
         // 渲染器
         keyTree.setCellRenderer(new KeyTreeCellRenderer());
@@ -147,7 +151,7 @@ public class KeyTreeDisplayPanel extends JPanel {
             }
         });
         JBScrollPane keyTreeScrollPane = new JBScrollPane(keyTree);
-        keyDisplayLoadingDecorator = new LoadingDecorator(keyTreeScrollPane, parent, 0);
+        keyDisplayLoadingDecorator = new LoadingDecorator(keyTreeScrollPane, project, 0);
 
         // toolbar
         final CommonActionsManager actionManager = CommonActionsManager.getInstance();
@@ -169,7 +173,7 @@ public class KeyTreeDisplayPanel extends JPanel {
                 .createActionToolbar(ActionPlaces.TOOLBAR, actions, true);
 
         // key分页panel
-        JPanel keyPagingPanel = createPagingPanel(parent);
+        JPanel keyPagingPanel = createPagingPanel();
 
         // key展示区域 包括key工具栏区域, key树状图区域, key分页区
         keyDisplayPanel = new JPanel(new BorderLayout());
@@ -180,6 +184,8 @@ public class KeyTreeDisplayPanel extends JPanel {
         keyDisplayPanel.add(keyPagingPanel, BorderLayout.SOUTH);
 
         splitterContainer.setFirstComponent(keyDisplayPanel);
+
+        renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), keys);
     }
 
     private static void updateTree(DefaultMutableTreeNode parentTargetNode, DefaultMutableTreeNode originalChildNode, String[] explodedKey, KeyInfo key) {
@@ -209,8 +215,8 @@ public class KeyTreeDisplayPanel extends JPanel {
             Object keyObj = currentChild.getUserObject();
             String nodeKey;
             // 中间节点
-            if (keyObj instanceof FragmentedKey) {
-                nodeKey = ((FragmentedKey) keyObj).getFragmentedKey();
+            if (keyObj instanceof FragmentedKey fragmentedKey) {
+                nodeKey = fragmentedKey.getFragmentedKey();
                 if (keyFragment.equals(nodeKey)) {
                     return currentChild;
                 }
@@ -229,65 +235,13 @@ public class KeyTreeDisplayPanel extends JPanel {
 
     /**
      * key分页panel
-     *
-     * @param parent
-     * @return
      */
     @NotNull
-    private JPanel createPagingPanel(ARedisKeyValueDisplayPanel parent) {
+    private JPanel createPagingPanel() {
         JPanel keyPagingPanel = new JPanel(new BorderLayout());
-
-        pageSizeLabel = new JBLabel("Page Size: " + currentPageKeys.size());
-        pageSizeLabel.setBorder(new EmptyBorder(0, 5, 0, 0));
+        pageSizeLabel = new JBLabel("Displayed key count: " + keys.size());
+        pageSizeLabel.setBorder(JBUI.Borders.emptyLeft(5));
         keyPagingPanel.add(pageSizeLabel, BorderLayout.NORTH);
-        JPanel pagePanel = new JPanel(new FlowLayout(FlowLayout.LEFT));
-        pageLabel = new JBLabel(String.format("Page %s of %s", pageIndex, getPageCount()));
-        pagePanel.add(pageLabel);
-
-        JButton prevPageButton = new JButton(AllIcons.Actions.Play_back);
-        JButton nextPageButton = new JButton(AllIcons.Actions.Play_forward);
-        pagePanel.add(prevPageButton);
-        pagePanel.add(nextPageButton);
-        keyPagingPanel.add(pagePanel, BorderLayout.SOUTH);
-
-        if (pageIndex <= 1) {
-            prevPageButton.setEnabled(false);
-        }
-        prevPageButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                if (pageIndex <= 1) {
-                    return;
-                }
-
-                updatePageDataOnPrevButtonClicked();
-                renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
-                nextPageButton.setEnabled(true);
-
-                if (pageIndex <= 1) {
-                    prevPageButton.setEnabled(false);
-                }
-            }
-        });
-
-        if (pageIndex >= getPageCount()) {
-            nextPageButton.setEnabled(false);
-        }
-        nextPageButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                if (pageIndex >= getPageCount()) {
-                    return;
-                }
-
-                updatePageDataOnNextButtonClicked();
-                renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
-                prevPageButton.setEnabled(true);
-                if (pageIndex >= getPageCount()) {
-                    nextPageButton.setEnabled(false);
-                }
-            }
-        });
         return keyPagingPanel;
     }
 
@@ -295,7 +249,7 @@ public class KeyTreeDisplayPanel extends JPanel {
      * 渲染keyTree
      */
     @SneakyThrows
-    public void renderKeyTree(String keyFilter, String groupSymbol) {
+    public void renderKeyTree(String keyFilter, String groupSymbol, List<String> preparedKeys) {
         keyDisplayLoadingDecorator.startLoading(false);
         ReadAction.nonBlocking(() -> {
             try {
@@ -305,35 +259,33 @@ public class KeyTreeDisplayPanel extends JPanel {
                 }
                 dbInfo.setKeyCount(dbSize);
                 flatRootNode = new DefaultMutableTreeNode(dbInfo);
-                // redis 查询前pageSize个key
-                allKeys = redisPoolManager.scan(SCAN_POINTER_START, keyFilter, pageSize, dbInfo.getIndex());
 
-                // exception occurred
-                if (allKeys == null) {
-                    return null;
+                // redis 查询前pageSize个key
+                if (preparedKeys == null) {
+                    keys = redisPoolManager.scanKeys(SCAN_POINTER_START, keyFilter, DEFAULT_PAGE_SIZE, dbInfo.getIndex());
+                    // exception occurred
+                    if (keys == null) {
+                        return null;
+                    }
+                } else {
+                    keys = preparedKeys;
                 }
 
-                if (CollectionUtils.isNotEmpty(allKeys)) {
-                    allKeys = allKeys.stream().sorted().collect(Collectors.toList());
-
-                    int size = allKeys.size();
-                    int start = (pageIndex - 1) * pageSize;
-                    int end = Math.min(start + pageSize, size);
-                    currentPageKeys = allKeys.subList(start, end);
-                    if (!CollectionUtils.isEmpty(currentPageKeys)) {
-                        for (String key : currentPageKeys) {
-                            DefaultMutableTreeNode keyNode = new DefaultMutableTreeNode(KeyInfo.builder()
-                                    .key(key)
-                                    .del(false)
-                                    .build());
-                            flatRootNode.add(keyNode);
-                        }
+                if (CollectionUtils.isNotEmpty(keys)) {
+                    for (String key : keys) {
+                        DefaultMutableTreeNode keyNode = new DefaultMutableTreeNode(KeyInfo.builder()
+                                .key(key)
+                                .del(false)
+                                .build());
+                        flatRootNode.add(keyNode);
                     }
                 }
 
-                updateKeyTree(groupSymbol);
-                updatePageLabel();
-                keyDisplayPanel.updateUI();
+                EventQueue.invokeLater(() -> {
+                    updateKeyTree(groupSymbol);
+                    updatePageLabel();
+                    keyDisplayPanel.updateUI();
+                });
             } finally {
                 keyDisplayLoadingDecorator.stopLoading();
             }
@@ -385,20 +337,14 @@ public class KeyTreeDisplayPanel extends JPanel {
     private ClearAction createClearAction() {
         ClearAction clearAction = new ClearAction();
         clearAction.setAction(e -> {
-            ConfirmDialog confirmDialog = new ConfirmDialog(
-                    project,
-                    "Confirm",
-                    "Are you sure you want to delete all the keys of the currently selected DB?",
-                    actionEvent -> {
-                        try (Jedis jedis = redisPoolManager.getJedis(dbInfo.getIndex())) {
-                            if (jedis != null) {
-                                jedis.flushDB();
-                            }
-                        }
-                        resetPageIndex();
-                        renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
-                    });
-            confirmDialog.show();
+            new FlushDbConfirmDialog(project, actionEvent -> {
+                try (Jedis jedis = redisPoolManager.getJedis(dbInfo.getIndex())) {
+                    if (jedis != null) {
+                        jedis.flushDB();
+                    }
+                }
+                renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), null);
+            }).show();
         });
         return clearAction;
     }
@@ -433,8 +379,7 @@ public class KeyTreeDisplayPanel extends JPanel {
                                 newKeyDialog.close(OK_EXIT_CODE);
                                 if (newKeyDialog.isReloadSelected()) {
                                     // 重新渲染keyTree
-                                    resetPageIndex();
-                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
+                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), null);
                                 }
                             }
                             break;
@@ -454,8 +399,7 @@ public class KeyTreeDisplayPanel extends JPanel {
                                 newKeyDialog.close(OK_EXIT_CODE);
                                 if (newKeyDialog.isReloadSelected()) {
                                     // 重新渲染keyTree
-                                    resetPageIndex();
-                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
+                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), null);
                                 }
                             }
                             break;
@@ -475,8 +419,7 @@ public class KeyTreeDisplayPanel extends JPanel {
                                 newKeyDialog.close(OK_EXIT_CODE);
                                 if (newKeyDialog.isReloadSelected()) {
                                     // 重新渲染keyTree
-                                    resetPageIndex();
-                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
+                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), null);
                                 }
                             }
                             break;
@@ -498,8 +441,7 @@ public class KeyTreeDisplayPanel extends JPanel {
                                 newKeyDialog.close(OK_EXIT_CODE);
                                 if (newKeyDialog.isReloadSelected()) {
                                     // 重新渲染keyTree
-                                    resetPageIndex();
-                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
+                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), null);
                                 }
                             }
                             break;
@@ -517,8 +459,7 @@ public class KeyTreeDisplayPanel extends JPanel {
                                 newKeyDialog.close(OK_EXIT_CODE);
                                 if (newKeyDialog.isReloadSelected()) {
                                     // 重新渲染keyTree
-                                    resetPageIndex();
-                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
+                                    renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), null);
                                 }
                             }
                     }
@@ -614,33 +555,13 @@ public class KeyTreeDisplayPanel extends JPanel {
     private RefreshAction createRefreshAction() {
         RefreshAction refreshAction = new RefreshAction();
         refreshAction.setAction(e -> {
-            renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol());
+            renderKeyTree(parent.getKeyFilter(), parent.getGroupSymbol(), null);
         });
         return refreshAction;
     }
 
-    private long getPageCount() {
-        int total = allKeys.size();
-        long result = total / pageSize;
-        long mod = total % pageSize;
-        return mod > 0 ? result + 1 : result;
-    }
-
-    private synchronized void updatePageDataOnPrevButtonClicked() {
-        pageIndex--;
-    }
-
-    private synchronized void updatePageDataOnNextButtonClicked() {
-        pageIndex++;
-    }
-
-    public void resetPageIndex() {
-        pageIndex = 1;
-    }
-
     private void updatePageLabel() {
-        pageLabel.setText(String.format("Page %s of %s", pageIndex, getPageCount()));
-        pageSizeLabel.setText("Page Size: " + currentPageKeys.size());
+        pageSizeLabel.setText("Displayed key count: " + keys.size());
     }
 
     /**
