@@ -103,7 +103,6 @@ public class ARedisKeyValueDisplayPanel extends JPanel implements Disposable {
                     dbInfo,
                     redisPoolManager,
                     this::renderValueDisplayPanel);
-            keyTreeDisplayPanel.renderKeyTree(this.getKeyFilter(), this.getGroupSymbol());
         } catch (RuntimeException e) {
             if ("exception occurred".equals(e.getMessage())) {
                 return;
@@ -140,8 +139,7 @@ public class ARedisKeyValueDisplayPanel extends JPanel implements Disposable {
                     } else {
                         searchTextField.addCurrentTextToHistory();
                     }
-                    keyTreeDisplayPanel.resetPageIndex();
-                    keyTreeDisplayPanel.renderKeyTree(getKeyFilter(), getGroupSymbol());
+                    keyTreeDisplayPanel.renderKeyTree(getKeyFilter(), getGroupSymbol(), null);
                 }
             }
         });
@@ -206,7 +204,7 @@ public class ARedisKeyValueDisplayPanel extends JPanel implements Disposable {
         valueDisplayScrollPanel.setHorizontalScrollBarPolicy(ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
         valueDisplayScrollPanel.setVerticalScrollBarPolicy(ScrollPaneConstants.VERTICAL_SCROLLBAR_NEVER);
 
-        LoadingDecorator loadingDecorator = new LoadingDecorator(valueDisplayScrollPanel, this, 0);
+        LoadingDecorator loadingDecorator = new LoadingDecorator(valueDisplayScrollPanel, project, 0);
         splitterContainer.setSecondComponent(loadingDecorator.getComponent());
         valueDisplayPanel.init(project, this, keyTreeDisplayPanel, key, redisPoolManager, dbInfo, loadingDecorator);
     }
