@@ -3,7 +3,12 @@ package com.mzyupc.aredis.view;
 import com.google.common.collect.Sets;
 import com.intellij.ide.CommonActionsManager;
 import com.intellij.ide.TreeExpander;
-import com.intellij.openapi.actionSystem.*;
+import com.intellij.openapi.Disposable;
+import com.intellij.openapi.actionSystem.ActionManager;
+import com.intellij.openapi.actionSystem.ActionPlaces;
+import com.intellij.openapi.actionSystem.ActionPopupMenu;
+import com.intellij.openapi.actionSystem.ActionToolbar;
+import com.intellij.openapi.actionSystem.DefaultActionGroup;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.application.ReadAction;
 import com.intellij.openapi.fileEditor.FileEditorManager;
@@ -12,7 +17,14 @@ import com.intellij.openapi.ui.LoadingDecorator;
 import com.intellij.ui.components.JBScrollPane;
 import com.intellij.ui.treeStructure.Tree;
 import com.intellij.util.messages.MessageBus;
-import com.mzyupc.aredis.action.*;
+import com.mzyupc.aredis.action.AddAction;
+import com.mzyupc.aredis.action.CloseAction;
+import com.mzyupc.aredis.action.ConsoleAction;
+import com.mzyupc.aredis.action.DeleteAction;
+import com.mzyupc.aredis.action.DuplicateAction;
+import com.mzyupc.aredis.action.EditAction;
+import com.mzyupc.aredis.action.InfoAction;
+import com.mzyupc.aredis.action.RefreshAction;
 import com.mzyupc.aredis.message.ARedisEventType;
 import com.mzyupc.aredis.message.ARedisStateChangeEvent;
 import com.mzyupc.aredis.message.ARedisStateChangeListener;
@@ -30,6 +42,7 @@ import com.mzyupc.aredis.view.editor.KeyValueDisplayVirtualFile;
 import com.mzyupc.aredis.view.render.ConnectionTreeCellRenderer;
 import com.mzyupc.aredis.vo.ConnectionInfo;
 import com.mzyupc.aredis.vo.DbInfo;
+import com.mzyupc.aredis.vo.Keyspace;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
 import org.jetbrains.annotations.Nullable;
@@ -42,8 +55,13 @@ import javax.swing.tree.TreePath;
 import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
-import java.util.*;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 import java.util.concurrent.CopyOnWriteArraySet;
 
 import static com.mzyupc.aredis.message.ARedisStateChangeListener.AREDIS_STATE_CHANGE_TOPIC;
@@ -52,7 +70,7 @@ import static com.mzyupc.aredis.utils.JTreeUtil.expandTree;
 /**
  * @author mzyupc@163.com
  */
-public class ConnectionManager {
+public class ConnectionManager implements Disposable {
     /**
      * connectionId-redisPoolManager
      */
@@ -138,8 +156,7 @@ public class ConnectionManager {
      */
     public Tree createConnectionTree(ARedisToolWindow parent, JPanel connectionPanel) {
         Tree connectionTree = new Tree();
-        // TODO 2022/4/16 开启链接的时候 因为读DB下的key数量导致整个软件卡住
-        connectionTreeLoadingDecorator = new LoadingDecorator(new JBScrollPane(connectionTree), parent, 0);
+        connectionTreeLoadingDecorator = new LoadingDecorator(new JBScrollPane(connectionTree), project, 0);
         connectionPanel.add(connectionTreeLoadingDecorator.getComponent(), BorderLayout.CENTER);
 //        connectionPanel.add(new JBScrollPane(connectionTree), BorderLayout.CENTER);
 
@@ -173,10 +190,12 @@ public class ConnectionManager {
                             ReadAction.nonBlocking(() -> {
                                 try {
                                     addDbs2Connection(connectionNode);
-                                    // 刷新节点
-                                    connectionTreeModel.reload(connectionNode);
-                                    // 展开当前连接
-                                    JTreeUtil.expandAll(connectionTree, new TreePath(new Object[]{connectionTreeRoot, connectionNode}), true);
+                                    EventQueue.invokeLater(() -> {
+                                        // 刷新节点
+                                        connectionTreeModel.reload(connectionNode);
+                                        // 展开当前连接
+                                        JTreeUtil.expandAll(connectionTree, new TreePath(new Object[]{connectionTreeRoot, connectionNode}), true);
+                                    });
                                 } finally {
                                     connectionTreeLoadingDecorator.stopLoading();
                                 }
@@ -198,13 +217,13 @@ public class ConnectionManager {
                         ReadAction.nonBlocking(() -> {
                             try {
                                 String connectionId = connectionInfo.getId();
-                                KeyValueDisplayVirtualFile keyValueDisplayVirtualFile = new KeyValueDisplayVirtualFile(connectionInfo.getName() + "-DB" + dbInfo.getIndex(),
-                                        project,
-                                        connectionInfo,
-                                        dbInfo,
-                                        connectionRedisMap.get(connectionId),
-                                        connectionManager);
                                 ApplicationManager.getApplication().invokeLater(() -> {
+                                    KeyValueDisplayVirtualFile keyValueDisplayVirtualFile = new KeyValueDisplayVirtualFile(connectionInfo.getName() + "-DB" + dbInfo.getIndex(),
+                                            project,
+                                            connectionInfo,
+                                            dbInfo,
+                                            connectionRedisMap.get(connectionId),
+                                            connectionManager);
                                     KeyValueDisplayFileSystem.getInstance(project).openEditor(keyValueDisplayVirtualFile);
                                     addEditorToMap(connectionId, keyValueDisplayVirtualFile);
                                 });
@@ -361,16 +380,18 @@ public class ConnectionManager {
 
         RedisPoolManager redisPoolManager = connectionRedisMap.get(connection.getId());
         int dbCount = redisPoolManager.getDbCount();
+        if (dbCount == 0) {
+            // return when exception
+            return;
+        }
+        Map<Integer, Keyspace> integerKeyspaceMap = redisPoolManager.infoKeyspace();
         propertyUtil.setDbCount(connection.getId(), dbCount);
 
         // 移除原有节点
         connectionNode.removeAllChildren();
         // 添加新节点
         for (int i = 0; i < dbCount; i++) {
-            Long keyCount = redisPoolManager.dbSize(i);
-            if (keyCount == null) {
-                return;
-            }
+            Long keyCount = Optional.ofNullable(integerKeyspaceMap.get(i)).map(Keyspace::getKeys).orElse(0L);
             DbInfo dbInfo = DbInfo.builder()
                     .index(i)
                     .keyCount(keyCount)
@@ -534,7 +555,9 @@ public class ConnectionManager {
                         }
                         DefaultMutableTreeNode connectionNode = (DefaultMutableTreeNode) selectionPath.getPath()[1];
                         addDbs2Connection(connectionNode);
-                        connectionTreeModel.reload(connectionNode);
+                        EventQueue.invokeLater(() -> {
+                            connectionTreeModel.reload(connectionNode);
+                        });
                     }
                 } finally {
                     connectionTreeLoadingDecorator.stopLoading();
@@ -681,6 +704,7 @@ public class ConnectionManager {
         return menu;
     }
 
+    @Override
     public void dispose() {
         for (RedisPoolManager redisPoolManager : connectionRedisMap.values()) {
             if (redisPoolManager != null) {
