@@ -2,13 +2,14 @@ package com.mzyupc.aredis.view.dialog;
 
 import com.intellij.openapi.fileTypes.PlainTextLanguage;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.ui.ComboBox;
 import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.ui.EditorTextField;
 import com.intellij.ui.components.JBLabel;
 import com.mzyupc.aredis.enums.RedisValueTypeEnum;
 import com.mzyupc.aredis.enums.ValueFormatEnum;
+import com.mzyupc.aredis.utils.DoubleDocument;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.batik.ext.swing.DoubleDocument;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -16,7 +17,6 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ItemEvent;
-import java.awt.event.ItemListener;
 import java.util.function.Consumer;
 
 import static com.mzyupc.aredis.view.textfield.EditorTextFieldManager.createEditorTextField;
@@ -68,9 +68,7 @@ public class AddRowDialog extends DialogWrapper {
 
     private JPanel createValuePanel() {
         switch (valueTypeEnum) {
-            case String:
-            case List:
-            case Set:
+            case String, List, Set:
                 return createSimpleValuePanel();
 
             case Zset:
@@ -127,14 +125,11 @@ public class AddRowDialog extends DialogWrapper {
     private JPanel createSimpleValuePanel() {
         valueTextArea = createEditorTextField(project, PlainTextLanguage.INSTANCE, "");
         JPanel stringTypePanel = new JPanel(new BorderLayout());
-        JComboBox<ValueFormatEnum> newKeyValueFormatEnumJComboBox = new JComboBox<>(ValueFormatEnum.values());
-        newKeyValueFormatEnumJComboBox.addItemListener(new ItemListener() {
-            @Override
-            public void itemStateChanged(ItemEvent e) {
-                if (ItemEvent.SELECTED == e.getStateChange()) {
-                    ValueFormatEnum formatEnum = (ValueFormatEnum) e.getItem();
-                    valueTextArea = formatValue(project, stringTypePanel, formatEnum, valueTextArea);
-                }
+        JComboBox<ValueFormatEnum> newKeyValueFormatEnumJComboBox = new ComboBox<>(ValueFormatEnum.values());
+        newKeyValueFormatEnumJComboBox.addItemListener(e -> {
+            if (ItemEvent.SELECTED == e.getStateChange()) {
+                ValueFormatEnum formatEnum = (ValueFormatEnum) e.getItem();
+                valueTextArea = formatValue(project, stringTypePanel, formatEnum, valueTextArea);
             }
         });
 
